♻️ refactor : communityDetail 페이지 fetch 함수 분리 적용
