--- src/types/index.ts
@@ -1,7 +1,0 @@
-export type commentData = {
-  id: number
-  name: string
-  date: string
-  content: string
-  imgUrl: string
-}
